Migrate from dep -> go mod
Since Go 1.11 there is a new way to structure packages, which don't
require the project to live under `$GOPATH`, and provides version
management.

`go mod` could superseed `dep`, given it is builtin on the core tooling
of the language and requires less dependencies to contribute.

This commit follows the [migration instructions] to adopt `go mod`.

The `README` has been updated, as well as `Makefile` and Travis
configuration.

The `main.go` has been moved into the root directory, so we are able to
`go build` and produce the `terraform-provider-zerotier` binary,
simplifying a lot the development.

[migration instructions]: https://github.com/golang/go/wiki/Modules#migrating-to-modules