Allow managing DIY Controllers
Zerotier provides a central controller on https://my.zerotier.com
and it is possible to run our own DIY controller if we desire as well.

Currently, this provider uses a hardcoded string pointing to the central
controller, which don't allow changes by downstream users.

This changes allow overriding the controller url on the provider itself,
with default falling back over to the central controller, including
documentation changes on `README`.

Built on top of https://github.com/cormacrelf/terraform-provider-zerotier/pull/8