Apply gofmt -s on the files
This commit applies `gofmt -s` which changes the code with a simplified
version if the formatter finds it is safe to apply.

The `make dev` command will apply the simplfied rewriting automatically
as well.

An extra change added was to move the `assignment_pool` resource schema
definition to a function, similar to how the `route` schema was defined,
to keep the consistency on the file.